created a main.py and a blockchain class that stores the blocks
